+++ test/not-ts.svelte
@@ -1,5 +1,0 @@
-<script>
-  let count = 1
-</script>
-
-<button on:click={() => count++}>{count}</button>