+++ test/src-not-found.svelte
@@ -1,3 +1,0 @@
-<script lang="ts" src="./not-found.js"></script>
-
-<div>Hello, world!</div>